requirements.txt
Flask==3.0.3
flask-cors==4.0.0
groq==0.8.0
httpx==0.27.2
python-dotenv==1.2.1